Bug: 825660 Reviewer: Peter Comment: Fix the case where an ad hoc Filr folder moved into a net folder then deleted can delete the entire net folder

+++ branches/Filr-1.0.1/ssf/main/src/org/kablink/teaming/module/binder/impl/AbstractBinderProcessor.java
@@ -518,8 +518,13 @@ public abstract class AbstractBinderProcessor extends CommonDependencyInjection
 					new String[] {binder.getPathName()});
 		}		
 		
-		if(binder.isMirrored()) { // The newly created binder is a mirrored one.
+		if (binder.isMirrored() || parent.isMirrored()) { 
+			// The newly created binder is a mirrored one or its parent is one.
 			// Make sure that the resource path we store is normalized.
+			if (!binder.isMirrored()) {
+				//If the binder is not mirrored, force it to be mirrored since only mirrored folders are allowed in a mirrored folder.
+				binder.setMirrored(true);
+			}
 	    	normalizeResourcePathIfInInput(binder, inputData);
 	    	binder.setResourceDriverName(parent.getResourceDriverName());
 						
@@ -1106,13 +1111,28 @@ public abstract class AbstractBinderProcessor extends CommonDependencyInjection
 							new String[] {binder.getPathName(), driver.getTitle()});
 				}
 				else {
-	    			ResourceSession session = getResourceDriverManager().getSession(driver, ResourceDriverManager.FileOperation.DELETE, binder.getParentBinder()).setPath(binder.getResourcePath());
-	    			try {
-	    				session.delete();
-	    			}
-	    			finally {
-	    				session.close();
-	    			}	
+					//Guard against deleting the whole mirrored source by accident
+					boolean okToDeleteSource = true;
+					if (binder.isAclExternallyControlled() && 
+							binder.getParentBinder() != null && binder.getParentBinder().isAclExternallyControlled() &&
+							binder.getParentBinder().getResourceDriverName().equals(binder.getResourceDriverName())) {
+						
+						//This is a sub-folder of a net folder. Check that it has a proper resource path
+						if (binder.getResourcePath() == null || binder.getResourcePath().equals("") || 
+								binder.getResourcePath().equals("/")) {
+							//Don't allow deleting of this source because it looks like the configuration wasn't properly completed.
+							okToDeleteSource = false;
+						}
+					}
+					if (okToDeleteSource) {
+		    			ResourceSession session = getResourceDriverManager().getSession(driver, ResourceDriverManager.FileOperation.DELETE, binder.getParentBinder()).setPath(binder.getResourcePath());
+		    			try {
+		    				session.delete();
+		    			}
+		    			finally {
+		    				session.close();
+		    			}	
+					}
 				}
     		}
     		catch(NotSupportedException e) {
@@ -1454,6 +1474,9 @@ public abstract class AbstractBinderProcessor extends CommonDependencyInjection
        Binder binder = null;
        try {
 			binder = addBinder(destination, sampleBinder.getEntryDef(), sampleBinder.getClass(), inputData, null, null);
+			//Also copy the configured definitions from the sample
+			binder.setDefinitions(sampleBinder.getDefinitions());
+			getCoreDao().flush();
        } catch (Exception e) {}
        
        return binder;
